+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This project uses [Semantic Versioning 2.0.0](http://semver.org/).
 
+## main
+
+- NEW: Added `registrar.get_domain_prices` to retrieve if domain is premium and prices to register, transfer, jjkjand renewal. (dnsimple/dnsimple-ruby#230)
+
 ## 5.2.0
 
 - CHANGED: `Certificate#expires_on` (date only) is deprecated in favor of `Certificate#expires_at` (timestamp). (dnsimple/dnsimple-ruby#190)